% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RForest_Model.R
\name{RForest_Model}
\alias{RForest_Model}
\title{Random Forest Model}
\usage{
RForest_Model(data, features, scaling, parameters)
}
\arguments{
\item{data}{The name of the dataset to be analyzed.}

\item{features}{The features to include in the analysis.}

\item{scaling}{The scaling scheme to be applied to the data.}

\item{parameters}{The parameters for the Random Forest model.}
}
\value{
a named list containing
}
\description{
Random Forest Model
}
\examples{
RForest_Model(data="Obesity Dataset", features=c(TRUE,TRUE,TRUE,TRUE,TRUE,FALSE,FALSE,FALSE,FALSE,
TRUE,TRUE,TRUE,TRUE,TRUE,TRUE), scaling="Robust Scaling", parameters=c(10, 3, 10, 10))
}
