% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyplot.R
\name{shinyplots}
\alias{shinyplots}
\title{shinyplot}
\usage{
shinyplots()
}
\value{
Shiny app
}
\description{
Calls Project 3 shiny app with descriptive plotting and widgets for MVR using RFR.
}
\examples{
\dontrun{ shinyplots()}
}
